refactor: consolidate menu building, fix MenuItemState bug and tag collisions
- Deduplicate build_menu_items (menu.rs) and recursive_build_nsmenu (events.rs) into single pub(crate) function in menu.rs
- Add MenuItemState handling (enabled/disabled/greyed) to build_menu_items, fixing bug where menu bar items always appeared enabled
- Fix tag collision between menu bar and context menus by using monotonically increasing counter in MenuState instead of max_key+1
- Standardize menu tag type on isize throughout (matching NSMenuItem API), removing i64/isize mismatch
Autoreview: dll/src/desktop/shell2/macos/menu.rs

+++ dll/src/desktop/shell2/macos/mod.rs
@@ -4609,7 +4609,7 @@ impl MacOSWindow {
         );
 
         // Look up callback from tag
-        let callback = match self.menu_state.get_callback_for_tag(tag as i64) {
+        let callback = match self.menu_state.get_callback_for_tag(tag) {
             Some(cb) => cb.clone(),
             None => {
                 log_warn!(
